Removed path manipulation for __base__ import.
Modifying sys.path is bad, but doing so within the package management
metadata script is double plus un-good.  (It’s also not needed for this
module, we can just import it relative.)  This isn’t fully armoured (it
assumes the current working directory is the same directory as
setup.py) but it’s a start.